requirements.txt
dropbox>=11.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-cloud-storage>=2.0.0
functions-framework>=3.0.0
python-dotenv>=1.0.0
